third_party/WebKit/LayoutTests/http/tests/cache/freshness-header-expected.txt

Resource-Freshness header is sent when a revalidation request is initiated by the Blink memory cache within the stale-while-revalidate window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resourceFreshnessHeader is "max-age=0,stale-while-revalidate=1,age=1"
PASS successfullyParsed is true

TEST COMPLETE
